+++ app/models/task.py
@@ -1,6 +1,6 @@
 import uuid
 
-from sqlalchemy import Column, String, Integer, ForeignKey, Table, DateTime
+from sqlalchemy import Column, String, Integer, ForeignKey, Table, DateTime, Float
 from sqlalchemy.orm import relationship
 from sqlalchemy.dialects.postgresql import UUID
 from app.db.database import Base
@@ -24,7 +24,7 @@ class Task(Base):
     programmings = relationship("Programming", secondary="programming_task", viewonly=True)
     programming_tasks = relationship("ProgrammingTask", back_populates="task", cascade="all, delete-orphan")
     people = Column(Integer, nullable=True)
-    performance = Column(Integer, nullable=True)
+    performance = Column(Float, nullable=True)
     material = Column(String, nullable=True)
     presentation = Column(String, nullable=True)
     fabricationCode = Column(String, nullable=True)
